Remove dead, unreachable rc yolo
